Use continuation object to thread context though handlers.
This replaces the argument list approach, which was verbose
and confusing, especially when looping and unpacking the
arguments, with an object with named properties.  It makes
the code much more readable and compact.

There seems to be little to no performance impact AS LONG
AS CONTINUATIONS ARE MONOMORPHIC.  Forgetting to set a
property on a continuation can lead to a major perf hit
on v8.  We may want to create an actual Continuation
constructor to help, but I don't feel the need just yet.

This change also has the nice side benefit of saving about
50 bytes from the min+gzip build.